+++ src/main/java/lotto/model/gameresult/GameResult.java
@@ -1,4 +1,6 @@
-package lotto.model;
+package lotto.model.gameresult;
+
+import lotto.model.Rank;
 
 import java.util.List;
