+++ src/main/java/lotto/model/lottos/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottos;
 
 import java.util.ArrayList;
 import java.util.List;
