+++ src/test/java/lotto/WinningLottoTest.java
@@ -1,8 +1,8 @@
 package lotto;
 
-import lotto.model.Lotto;
-import lotto.model.LottoNumber;
-import lotto.model.WinningLotto;
+import lotto.model.lottos.Lotto;
+import lotto.model.lottos.LottoNumber;
+import lotto.model.winninglotto.WinningLotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
